// mobile/src/api/generated/models/BlockedUserDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GrannFix API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BlockedUserDto
 */
export interface BlockedUserDto {
    /**
     * 
     * @type {string}
     * @memberof BlockedUserDto
     */
    userId?: string;
    /**
     * 
     * @type {string}
     * @memberof BlockedUserDto
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof BlockedUserDto
     */
    profileImageUrl?: string;
    /**
     * 
     * @type {Date}
     * @memberof BlockedUserDto
     */
    blockedAt?: Date;
}

/**
 * Check if a given object implements the BlockedUserDto interface.
 */
export function instanceOfBlockedUserDto(value: object): value is BlockedUserDto {
    return true;
}

export function BlockedUserDtoFromJSON(json: any): BlockedUserDto {
    return BlockedUserDtoFromJSONTyped(json, false);
}

export function BlockedUserDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): BlockedUserDto {
    if (json == null) {
        return json;
    }
    return {
        
        'userId': json['userId'] == null ? undefined : json['userId'],
        'name': json['name'] == null ? undefined : json['name'],
        'profileImageUrl': json['profileImageUrl'] == null ? undefined : json['profileImageUrl'],
        'blockedAt': json['blockedAt'] == null ? undefined : (new Date(json['blockedAt'])),
    };
}

export function BlockedUserDtoToJSON(json: any): BlockedUserDto {
    return BlockedUserDtoToJSONTyped(json, false);
}

export function BlockedUserDtoToJSONTyped(value?: BlockedUserDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'userId': value['userId'],
        'name': value['name'],
        'profileImageUrl': value['profileImageUrl'],
        'blockedAt': value['blockedAt'] == null ? value['blockedAt'] : value['blockedAt'].toISOString(),
    };
}